Rewrite based on a moose object hierarchy
The interface changed (and will probably change again)
But, the test series pass as before.

I kept the logic for date, but I'm not really pleased about it.
Perhaps, the perl module should be more strict and impose a 'good' date
(and less strict to not check if the date really exists...)